backend/requirements.txt
absl-py==0.13.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.4
astroid==2.6.4
billiard==3.3.0.23
celery==3.1.18
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==3.2
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
kombu==3.0.37
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
ortools==9.0.9048
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.4
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
redis==3.2.0
requests==2.26.0
sentry-sdk==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
vine==1.3.0